Moved Direction to own module
